move grpc client to own mod
