+++ wavesexchange_apis/src/lib.rs
@@ -1,4 +1,4 @@
-mod client;
+mod clients;
 mod error;
 #[cfg(test)]
 mod tests;
@@ -9,5 +9,8 @@ pub mod models;
 #[macro_use]
 extern crate async_trait;
 
-pub use client::{ApiBaseUrl, HttpClient};
+pub use clients::{
+    grpc::GrpcClient,
+    http::{ApiBaseUrl, HttpClient},
+};
 pub use error::Error;
